apps/api/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.7.14
chardet==5.2.0
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
deprecation==2.1.0
fastapi==0.116.1
gitdb==4.0.12
GitPython==3.1.45
google-genai
gotrue==2.12.3
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10

packaging==25.0
postgrest==1.1.1
prompt_toolkit==3.0.51
pydantic==2.11.7
pydantic_core==2.33.2
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-magic==0.4.27
python-multipart==0.0.20
PyYAML==6.0.2
firecrawl-py==1.5.0
tweepy==4.14.0
realtime==2.6.0

repo2text @ git+https://github.com/CodeGuide-dev/repo2text.git@fdf67a8416a920747fac34e1398421124365ffdb
requests==2.32.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.47.2
storage3==0.12.0
StrEnum==0.4.15
supabase==2.17.0
supafunc==0.10.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
playwright==1.51.0
Pillow==11.1.0
nest_asyncio==1.6.0
markdown==3.7

watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
